+++ src/lib/axios/setupDeviceInterceptors.ts
@@ -1,114 +1,7 @@
-/* eslint-disable no-param-reassign */
-
 import { Mutex } from "async-mutex";
 import { AxiosInstance } from "axios";
 import makeSignature from "@/utils/make-signature";
-import { renewToken } from "../api/auth.api";
 import { getDecryptedItem } from "../auth/secureStorage";
-import { deleteCookie, getToken, setCookie } from "../cookies";
-import { getClientCookie } from "../cookies/client";
-
-let isRefreshing = false;
-let refreshPromise: Promise<any> | null = null;
-const mutex = new Mutex();
-
-export const setupInterceptors = (axiosInstance: AxiosInstance) => {
-  axiosInstance.interceptors.request.use(
-    async (config) => {
-      let token;
-
-      if (typeof window === "undefined") {
-        token = await getToken("accessToken");
-      } else {
-        token = getClientCookie("accessToken");
-      }
-
-      if (token) {
-        config.headers.Authorization = `Bearer ${token}`;
-      }
-      return config;
-    },
-    (error) => Promise.reject(error)
-  );
-
-  // NOTE - 응답 인터셉터
-  axiosInstance.interceptors.response.use(
-    (response) => response,
-    async (error) => {
-      const e = error as any;
-      const originalRequest = error.config as any;
-
-      if (error.response?.status === 403) {
-        return Promise.reject(error);
-      }
-
-      if (error.response?.status === 400) {
-        // eslint-disable-next-line no-console
-        console.error(e.response.data.message);
-        return Promise.reject(error);
-      }
-
-      if (error.response?.status === 401 && !originalRequest.retryFlag) {
-        originalRequest.retryFlag = true;
-
-        // 갱신 중이면 기다림
-        if (isRefreshing && refreshPromise) {
-          await refreshPromise;
-          const token = await getToken("accessToken");
-          originalRequest.headers.Authorization = `Bearer ${token}`;
-          return axiosInstance(originalRequest);
-        }
-
-        // 갱신 시작
-        isRefreshing = true;
-        refreshPromise = mutex.runExclusive(async () => {
-          try {
-            const refreshToken = await getToken("refreshToken");
-
-            if (process.env.NODE_ENV === "development") {
-              // eslint-disable-next-line no-console
-              console.log(`refresh ${refreshToken}`);
-            }
-
-            if (!refreshToken) {
-              window.location.href = "/login";
-              throw error;
-            }
-
-            const { accessToken } = await renewToken({ refreshToken });
-
-            if (process.env.NODE_ENV === "development") {
-              // eslint-disable-next-line no-console
-              console.log("refresh: success ✅");
-            }
-
-            await setCookie("accessToken", accessToken);
-            return accessToken;
-          } catch (refreshError: any) {
-            const code = refreshError?.response?.data?.code;
-            if (["FORBIDDEN", "UNAUTHORIZED"].includes(code)) {
-              window.location.href = "/login";
-            } else {
-              await deleteCookie("accessToken");
-              await deleteCookie("refreshToken");
-              window.location.href = "/login";
-            }
-            throw refreshError;
-          } finally {
-            isRefreshing = false;
-            refreshPromise = null;
-          }
-        });
-
-        const accessToken = await refreshPromise;
-        originalRequest.headers.Authorization = `Bearer ${accessToken}`;
-        return axiosInstance(originalRequest);
-      }
-
-      return Promise.reject(error);
-    }
-  );
-};
 
 type CacheKey = `${string}:${string}:${string}`;
 
@@ -131,6 +24,7 @@ export const setupDeviceInterceptors = (axiosInstance: AxiosInstance) => {
     const searchParams = new URLSearchParams(config.params).toString();
     const uri = searchParams ? `${rawUrl}?${searchParams}` : rawUrl;
 
+    // eslint-disable-next-line
     await signatureMutex.runExclusive(async () => {
       const key = getCacheKey(method, `/v1${uri}`, "POS");
       const cached = signatureCache[key];
@@ -153,7 +47,7 @@ export const setupDeviceInterceptors = (axiosInstance: AxiosInstance) => {
         if (!deviceInfo || !secretKey) {
           // eslint-disable-next-line no-console
           console.warn("❌ deviceInfo 또는 secretKey가 없습니다.");
-          return;
+          return config;
         }
 
         const timestamp = Date.now().toString();
@@ -186,9 +80,15 @@ export const setupDeviceInterceptors = (axiosInstance: AxiosInstance) => {
           });
         }
 
-        config.headers["x-ew-access-key"] = deviceInfo.deviceId;
-        config.headers["x-ew-signature"] = signature;
-        config.headers["x-ew-timestamp"] = timestamp;
+        return {
+          ...config,
+          headers: {
+            ...(config.headers || {}),
+            "x-ew-access-key": deviceInfo.deviceId,
+            "x-ew-signature": signature,
+            "x-ew-timestamp": timestamp,
+          },
+        };
       }
     });
